requirements.txt
-e git+https://github.com/JacobPfau/procgenAISC.git@7821f2c00be9a4ff753c6d54b20aed26028ca812#egg=procgen&subdirectory=../../procgenAISC
-e git+https://github.com/modanesh/cliport.git@ec9d1c96e0fd6d8f2028d743d4107ad3081c0c6b#egg=cliport&subdirectory=../../cliport
absl-py==0.7.0
antlr4-python3-runtime==4.8
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==1.6.0
configparser==7.0.0
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
ftfy==6.2.0
future==1.0.0
gdown==3.13.1
gitdb==4.0.11
GitPython==3.1.43
glcontext==2.5.0
glfw==1.12.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.63.0
gym==0.17.3
gym3==0.3.3
hydra-core==1.0.5
idna==3.7
imageio==2.9.0
imageio-ffmpeg==0.3.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
kornia==0.4.1
llvmlite==0.41.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
meshcat==0.0.18
moderngl==5.10.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.22.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.0.6
opencv-python==4.1.2.30
packaging==21.3
pandas==2.0.3
pathtools==0.1.2
pillow==10.3.0
platformdirs==4.2.2
promise==2.3
protobuf==3.20.1
psutil==5.9.8
pyasn1==0.6.0
pyasn1-modules==0.4.0
pybullet==3.0.4
pycparser==2.22
pyglet==1.5.0
pyod==2.0.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.0.3
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sacremoses==0.1.1
scikit-image==0.17.2
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==2.1.1
setproctitle==1.3.3
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
subprocess32==3.5.4
sympy==1.12
tdqm==0.0.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2023.7.10
tokenizers==0.10.3
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.4
transformers==4.3.2
transforms3d==0.3.1
triton==2.1.0
typing-extensions==4.11.0
tzdata==2024.1
u-msgpack-python==2.8.0
urllib3==2.2.1
wandb==0.17.3
watchdog==0.10.4
wcwidth==0.2.13
werkzeug==3.0.3
zipp==3.18.1
